requirements.txt
# CUDA 12.8 stable index
--extra-index-url https://download.pytorch.org/whl/cu128

# Core DL stack - Synchronized for RTX 50-series (sm_120)
torch==2.11.0+cu128
torchvision==0.26.0+cu128
torchaudio==2.11.0+cu128

# TensorRT - Pinning to 10.x stable
torch-tensorrt==2.11.0+cu128
tensorrt==10.15.1.29
tensorrt-cu12==10.15.1.29

# OpenCOOD core
cmake>=3.26
easydict==1.13
numpy==1.26.4
scipy==1.13.1
numba==0.60.0
cython==0.29.37
opencv-python==4.10.0.84
scikit-image==0.24.0
open3d==0.18.0
matplotlib==3.8.4
tqdm==4.67.1
PyYAML==6.0.2
tensorboardX==2.6.2.2
shapely==2.0.6
einops==0.8.1
timm==1.0.15
scikit-learn==1.5.2
nvidia-modelopt==0.43.0

# Sparse Engine - Unified on CUDA 12.6 binaries (compatible with 12.8 runtime)
# REMOVED: cumm-cu120 (from your pip list) to prevent collision
cumm-cu126==0.7.11
spconv-cu126==2.3.8